refactoring: #845 add and use convenience import for module "coordinate_reference_systems"
