+++ tests/conversion/conftest.py
@@ -7,7 +7,7 @@ from django.conf import settings
 
 from osmaxx.conversion import output_format, status
 from osmaxx.conversion.converters.converter_gis import detail_levels
-from osmaxx.conversion.constants import coordinate_reference_systems as crs
+from osmaxx.conversion import coordinate_reference_systems as crs
 
 format_list = output_format.DEFINITIONS.keys()
